projects/family/requirements.txt
Django>1.8,<1.9
simplejson
django-colorful
git+https://github.com/bobah101/yandex-maps.git
django-extensions
django-easy-select2
pytils
django-bootstrap3
django-qrcode
xhtml2pdf
django-grappelli
django-celery
redis==2.10.5
django-tinymce==2.2.0
Pillow==2.6.1
django-imagekit==3.2.6
django-formtools==1.0
django-messages
